# openapi.yaml
openapi: 3.0.3
info:
  title: LeadCatch CRM API
  description: |
    LeadCatch is a powerful CRM backend for lead capture, enrichment, and scoring.

    ## Features
    - Lead CRUD operations
    - AI-powered lead parsing and enrichment
    - Waterfall enrichment (Apollo, Clearbit, Hunter.io, Snov.io)
    - Lead scoring and auto-tagging
    - Duplicate detection
    - CSV import/export
    - Webhook notifications
    - Background job processing

  version: 1.0.0
  contact:
    name: Ahmed Mahmoud
    email: ahmed@leadcatch.io
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.leadcatch.io
    description: Production server
...
paths:
...
  /api/leads:
    post:
      summary: Create Lead Manually
      tags: [Leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
      responses:
        "201":
          description: Lead created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/leads/{userId}:
    get:
      summary: Get All Leads for User
      tags: [Leads]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, interested, closed]
        - name: platform
          in: query
          schema:
            type: string
        - name: minScore
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /api/leads/{userId}/{leadId}:
    get:
      summary: Get Single Lead
      tags: [Leads]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/leads/{leadId}:
    patch:
      summary: Update Lead
      tags: [Leads]
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadRequest"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete Lead
      tags: [Leads]
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Lead deleted successfully
  /api/leads/parse:
    post:
      summary: Parse Text with AI
      tags: [Leads]
      description: Extract lead information from free-form text using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, text]
              properties:
                userId:
                  type: string
                  format: uuid
                text:
                  type: string
                  description: Free-form text to parse (email signature, business card, etc.)
      responses:
        "201":
          description: Lead parsed and created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
  /api/leads/import:
    post:
      summary: Import Leads from CSV
      tags: [Leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, leads]
              properties:
                userId:
                  type: string
                  format: uuid
                leads:
                  type: array
                  items:
                    $ref: "#/components/schemas/ImportLead"
      responses:
        "201":
          description: Leads imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  leads:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
  /api/leads/{leadId}/enrich:
    post:
      summary: Enrich Lead Manually
      tags: [Enrichment]
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lead enriched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
  /api/leads/{leadId}/score:
    post:
      summary: Calculate Lead Score
      tags: [Enrichment]
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Score calculated
          content:
            application/json:
              schema:
                type: object
                properties:
                  leadId:
                    type: string
                    format: uuid
                  score:
                    type: integer
                    minimum: 0
                    maximum: 100
  /api/leads/{leadId}/auto-tag:
    post:
      summary: Auto-tag Lead with AI
      tags: [Enrichment]
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tags generated and assigned
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
  /api/leads/dedupe:
    post:
      summary: Find Duplicate Leads
      tags: [Leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Duplicates found
          content:
            application/json:
              schema:
                type: object
                properties:
                  duplicates:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
  /api/leads/{userId}/export:
    get:
      summary: Export Leads to CSV
      tags: [Leads]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /api/leads/bulk:
    post:
      summary: Bulk Operations
      tags: [Leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, leadIds]
              properties:
                action:
                  type: string
                  enum: [delete, update]
                leadIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                updates:
                  type: object
                  description: Required for update action
      responses:
        "200":
          description: Bulk operation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/leads/search:
    get:
      summary: Search Leads
      tags: [Leads]
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /api/leads/batch-enrich:
    post:
      summary: Batch Enrich Leads
      tags: [Enrichment]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                leadIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                limit:
                  type: integer
                  default: 50
      responses:
        "200":
          description: Batch enrichment results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchEnrichmentResult"
  /api/scrape:
    post:
      summary: Scrape LinkedIn Profile
      tags: [Enrichment]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, targetUrl]
              properties:
                userId:
                  type: string
                  format: uuid
                targetUrl:
                  type: string
                  format: uri
                platform:
                  type: string
                  default: linkedin
      responses:
        "200":
          description: Scrape and enrichment successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
  /api/leads/{leadId}/notes:
    post:
      summary: Add Note to Lead
      tags: [Notes]
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, content]
              properties:
                userId:
                  type: string
                  format: uuid
                content:
                  type: string
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                type: object
                properties:
                  note:
                    $ref: "#/components/schemas/Note"

    get:
      summary: Get Notes for Lead
      tags: [Notes]
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
  /api/notes/{noteId}:
    delete:
      summary: Delete Note
      tags: [Notes]
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Note deleted
...
  /api/tags/{userId}:
    get:
      summary: Get User Tags
      tags: [Tags]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
...
  /api/webhooks/{webhookId}:
    delete:
      summary: Delete Webhook
      tags: [Webhooks]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook deleted
...
components:
  schemas:
...
    CreateLeadRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        job_title:
          type: string
        company:
          type: string
        location:
          type: string
        linkedin_url:
          type: string
          format: uri
...
    LeadResponse:
      type: object
      properties:
        message:
          type: string
        lead:
          $ref: "#/components/schemas/Lead"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...